wallet-overview-api/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-binance
bcrypt